.......... [DEV-3693] fixed style issues

+++ src/libs/zbxxml/traverse.c
@@ -148,7 +148,7 @@ int	zbx_xml_traverse(char **data, char *error, int maxerrlen, zbx_xml_resolv_fun
 
 	zbx_free(*data);
 	*data = zbx_malloc(NULL, (size_t)(size + 1));
-	memcpy(*data, (const char *)mem, size + 1);
+	memcpy(*data, (const char *)mem, (size_t)(size + 1));
 	xmlFree(mem);
 	ret = SUCCEED;
 clean:
